+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `--output` to allow writing to files, `--no-confirm-overwrite` to allow binary to overwrite existing files.
-
+### Changed
+- Logs are now printed to stderr instead of stdout
 
 ## [0.3.3] - 2019-05-23